+++ opentelemetry-otlp/src/lib.rs
@@ -366,10 +366,6 @@ pub enum Error {
     #[error("the lock of the {0} has been poisoned")]
     PoisonedLock(&'static str),
 
-    /// The pipeline will need a exporter to complete setup. Throw this error if none is provided.
-    #[error("no exporter builder is provided, please provide one using with_exporter() method")]
-    NoExporterBuilder,
-
     /// Unsupported compression algorithm.
     #[error("unsupported compression algorithm '{0}'")]
     UnsupportedCompressionAlgorithm(String),
@@ -408,3 +404,8 @@ pub enum Protocol {
     /// HTTP protocol with binary protobuf
     HttpBinary,
 }
+
+#[derive(Debug, Default)]
+#[doc(hidden)]
+/// Placeholder type when no exporter pipeline has been configured in telemetry pipeline.
+pub struct NoExporterConfig(());