+++ opentelemetry-sdk/src/attributes/set.rs
@@ -104,7 +104,7 @@ impl Eq for HashKeyValue {}
 ///
 /// This must implement [Hash], [PartialEq], and [Eq] so it may be used as
 /// HashMap keys and other de-duplication methods.
-#[derive(Clone, Debug, Hash, PartialEq, Eq)]
+#[derive(Clone, Default, Debug, Hash, PartialEq, Eq)]
 pub struct AttributeSet(BTreeSet<HashKeyValue>);
 
 impl From<&[KeyValue]> for AttributeSet {