+++ opentelemetry/src/context.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "trace")]
+use crate::trace::context::SynchronizedSpan;
 use std::any::{Any, TypeId};
 use std::cell::RefCell;
 use std::collections::HashMap;
@@ -74,6 +76,8 @@ thread_local! {
 /// ```
 #[derive(Clone, Default)]
 pub struct Context {
+    #[cfg(feature = "trace")]
+    pub(super) span: Option<Arc<SynchronizedSpan>>,
     entries: HashMap<TypeId, Arc<dyn Any + Sync + Send>, BuildHasherDefault<IdHasher>>,
 }
 
@@ -301,6 +305,22 @@ impl Context {
         ContextGuard {
             previous_cx,
             _marker: PhantomData,
+        }
+    }
+
+    #[cfg(feature = "trace")]
+    pub(super) fn current_with_synchronized_span(value: SynchronizedSpan) -> Self {
+        Context {
+            span: Some(Arc::new(value)),
+            entries: Context::map_current(|cx| cx.entries.clone()),
+        }
+    }
+
+    #[cfg(feature = "trace")]
+    pub(super) fn with_synchronized_span(&self, value: SynchronizedSpan) -> Self {
+        Context {
+            span: Some(Arc::new(value)),
+            entries: self.entries.clone(),
         }
     }
 }